updating from base

+++ test/perf2/go_benchmark.jl
@@ -1,6 +1,8 @@
 # Benchmark implementing the board logic for the game of go and
 # exercising it by playing random games. Derived from
 # http://www.lysator.liu.se/~gunnar/gtp/brown-1.0.tar.gz
+import Base.ref
+
 const EMPTY = 0
 const WHITE = 1
 const BLACK = 2
@@ -40,7 +42,7 @@ neighbor(i::Int, j::Int, k::Int) = (i + deltai[k], j + deltaj[k])
 
 type Board
   size::Int
-  komi::Float
+  komi::Float64
   
   # Board represented by a 1D array. The first board_size*board_size
   # elements are used. Vertices are indexed row by row, starting with 0
@@ -86,7 +88,7 @@ ref(board::Board, i::Int, j::Int) = board.board[i, j]
 POS(board::Board, i::Int, j::Int) = (j - 1) * board.size + i
 IJ(board::Board, pos::Int) = (1 + mod((pos - 1), board.size), 1 + fld(pos - 1, board.size))
 
-function set_komi(board::Board, komi::Float)
+function set_komi(board::Board, komi::Float64)
   board.komi = komi
 end
